Reject IPTC extended record lengths beyond the remaining data
A file-controlled extended record length of 0x7FFFFFFF overflowed the
index on the next loop iteration and threw a raw
IndexOutOfBoundsException. ByteArrayExtensions.slice now computes the
end index in Long space so wrapped-around ranges cannot hide the error.

+++ src/commonMain/kotlin/de/stefan_oltmann/kim/common/ByteArrayExtensions.kt
@@ -107,7 +107,13 @@ internal fun ByteArray.getRemainingBytes(startIndex: Int): ByteArray {
 }
 
 internal fun ByteArray.slice(startIndex: Int, count: Int): ByteArray {
-    val endIndex = (startIndex + count).coerceAtMost(size)
+
+    /*
+     * Long arithmetic so a file-controlled count cannot overflow into a
+     * wrapped-around range that would silently hide the wrong bytes.
+     */
+    val endIndex = (startIndex.toLong() + count).coerceAtMost(size.toLong()).toInt()
+
     return sliceArray(startIndex until endIndex)
 }
